tests/requirements.txt
pytest==3.1.3
pytest-asyncio==0.5.0
pytest-cov==2.5.1
pytest-aiohttp==0.1.3